@@ -10,12 +10,12 @@
     "check": "svelte-check --tsconfig ./tsconfig.json --compiler-warnings 'a11y-no-static-element-interactions:ignore,a11y-no-noninteractive-element-interactions:ignore,a11y-click-events-have-key-events:ignore'"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^2.4.1",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tsconfig/svelte": "^4.0.0",
-    "svelte": "^4.0.0",
-    "svelte-check": "^3.4.3",
-    "svelte-preprocess": "^5.0.3",
-    "typescript": "^5.0.0",
-    "vite": "^4.0.0"
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
+    "svelte-preprocess": "^6.0.0",
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4"
   }
 }
