@@ -1,6 +1,8 @@
 <!-- SPDX-License-Identifier: AGPL-3.0-or-later -->
 <!-- SPDX-FileCopyrightText: 2022 Jani Nikula <jani@nikula.org> -->
 <script lang='ts'>
+  import { stopPropagation } from 'svelte/legacy';
+
   import { Fullscreen } from './lib/Fullscreen';
   import * as timeutil from './lib/time-util';
   import Ball from './lib/Ball.svelte';
@@ -19,15 +21,15 @@
   }
 
   // Hack to "live update" generic stuff once per second
-  let __counter = 0;
+  let __counter = $state(0);
   setInterval(() => __counter++, 1000)
 
-  $: live_update = function(thing: string): string {
+  let live_update = $derived(function(thing: string): string {
     // Do something with __counter to react to changes
     let dummy = __counter;
     dummy = dummy
     return thing;
-  }
+  })
 
   // ui pages
   const UiPage = {
@@ -36,7 +38,7 @@
     EDIT: 2,
   };
 
-  let ui_page = UiPage.START;
+  let ui_page = $state(UiPage.START);
 
   function ui_load_game(save_game: SaveGameId): void {
     if (!save_game.timestamp)
@@ -244,8 +246,8 @@
     }
   }
 
-  let show_menu = false;
-  let show_stats = false;
+  let show_menu = $state(false);
+  let show_stats = $state(false);
 
   function ui_show_menu() {
     show_menu = true;
@@ -272,14 +274,14 @@
 
 </script>
 
-<svelte:window on:keydown={ui_key_down} />
+<svelte:window onkeydown={ui_key_down} />
 
 <main>
-  {#if ui_page == UiPage.START }
+  {#if ui_page == UiPage.START}
 
     <div class='grid-container'>
-      <div class='name-input-card middle {$names.can_new_game() ? "" : "unavailable"}' on:click={ui_new_game}>
-	<div class='info-card-copyright' on:click|stopPropagation={() => false}><a href="https://groovescore.app">&copy; 2022-2024 Jani Nikula<br>License: AGPL 3.0 or later &#x1f517;</a></div>
+      <div class='name-input-card middle {$names.can_new_game() ? "" : "unavailable"}' onclick={ui_new_game}>
+	<div class='info-card-copyright' onclick={stopPropagation(() => false)}><a href="https://groovescore.app">&copy; 2022-2024 Jani Nikula<br>License: AGPL 3.0 or later &#x1f517;</a></div>
 	<div></div>
 	<div>GrooveScore</div>
 	<div>Snooker</div>
@@ -289,7 +291,7 @@
       </div>
       {#each $names.names as player_name (player_name.id)}
 	<div class='name-input-card {ui_name_input_card_style(player_name)}'>
-	  <input class='name-input' size=22 minlength=1 maxlength=22 placeholder='enter name' bind:value='{player_name.name}' on:click|stopPropagation={() => {}}/>
+	  <input class='name-input' size=22 minlength=1 maxlength=22 placeholder='enter name' bind:value='{player_name.name}' onclick={stopPropagation(() => {})}/>
 	  <div></div>
 	  <div></div>
 	  <div></div>
@@ -300,10 +302,10 @@
       {/each}
 
       {#each $game.saved_games as save_game, index (save_game.slot) }
-	<div class='info-card {save_game.timestamp ? "" : "unavailable"}' on:click={() => ui_load_game(save_game)}>
+	<div class='info-card {save_game.timestamp ? "" : "unavailable"}' onclick={() => ui_load_game(save_game)}>
 	  <div>Game save {index}</div>
 	  <div></div>
-	  {#if save_game.timestamp }
+	  {#if save_game.timestamp}
 	    <div>Started</div>
 	    <div>{timeutil.format_date(save_game.timestamp)}</div>
 	    <div>{timeutil.format_time(save_game.timestamp)}</div>
@@ -319,9 +321,9 @@
 
     </div>
 
-  {:else if ui_page == UiPage.PLAY }
+  {:else if ui_page == UiPage.PLAY}
     <div class='grid-container'>
-      <div class='score-card middle' on:click={ui_show_menu}>
+      <div class='score-card middle' onclick={ui_show_menu}>
 	<div>{ live_update($game.state.get_frame_time()) }</div>
 	<div>Frames ({$game.state.num_frames})</div>
 	<div>
@@ -329,7 +331,7 @@
 	  <div>(Remaining {$game.state.num_points()})</div>
 	</div>
 	<div>Break</div>
-	{#if $game.state.respot_black }
+	{#if $game.state.respot_black}
 	  <div class='highlight'>re-spot black</div>
 	{:else}
 	  <div></div>
@@ -337,7 +339,7 @@
 	<div class='card-button'>Menu</div>
       </div>
       {#each $game.state.get_players() as player (player.pid)}
-	<div class='score-card {ui_score_card_player_style(player)}' on:click={() => ui_click_player(player)}>
+	<div class='score-card {ui_score_card_player_style(player)}' onclick={() => ui_click_player(player)}>
 	  <div>{player.name}</div>
 	  <div>{player.frame_wins}</div>
 	  <div class='score-card-points'>{player.points}</div>
@@ -348,9 +350,9 @@
 	    <div>({player.last_break})</div>
 	    <div class='score-card-break unavailable'><Break balls={player._last_break}></Break></div>
 	  {/if}
-	  {#if $game.state.is_current_player(player.pid) && $game.state.can_end_turn() }
+	  {#if $game.state.is_current_player(player.pid) && $game.state.can_end_turn()}
 	    <div title='Shortcut: [space]' class='card-button'>End turn</div>
-	  {:else if $game.state.is_winner(player.pid) }
+	  {:else if $game.state.is_winner(player.pid)}
 	    <div class='highlight'>Frame Winner</div>
 	  {:else}
 	    <div></div>
@@ -400,7 +402,7 @@
     </div>
   {:else}
     <div class='grid-container'>
-      <div class='score-card middle' on:click={ui_goto_play_page}>
+      <div class='score-card middle' onclick={ui_goto_play_page}>
 	<div>{ live_update($game.state.get_frame_time()) }</div>
 	<div>Frames ({$game.state.num_frames})</div>
 	<div>
@@ -424,8 +426,8 @@
 	    <div class='score-card-break unavailable'><Break balls={player._last_break}></Break></div>
 	  {/if}
 	  <div class='double-button'>
-	    <div class='card-button' on:click={() => ui_player_edit_points(player.pid, -1)}>&ndash;</div>
-	    <div class='card-button' on:click={() => ui_player_edit_points(player.pid, 1)}>+</div>
+	    <div class='card-button' onclick={() => ui_player_edit_points(player.pid, -1)}>&ndash;</div>
+	    <div class='card-button' onclick={() => ui_player_edit_points(player.pid, 1)}>+</div>
 	  </div>
 	</div>
       {/each}
@@ -475,19 +477,19 @@
 
   {/if}
 
-{#if ui_page != UiPage.START }
+{#if ui_page != UiPage.START}
   <Dialog bind:show={show_menu}>
     <div class='dialog'>
       <div class='menu'>
 	<div class='menu-column'>
-	  <div title='Shortcut: s' class='menu-button' on:click={ui_show_stats}>Statistics</div>
-	  <div title='Shortcut: e' class='menu-button' on:click={ui_goto_edit_page}>Edit</div>
-	  <div title='Shortcut: FIXME' class='menu-button {$game.state.can_end_frame() ? "" : "unavailable"}' on:click={ui_end_frame}>End frame</div>
-	  <div title='Shortcut: FIXME' class='menu-button {$game.state.can_new_frame() ? "" : "unavailable"}' on:click={ui_new_frame}>New frame</div>
+	  <div title='Shortcut: s' class='menu-button' onclick={ui_show_stats}>Statistics</div>
+	  <div title='Shortcut: e' class='menu-button' onclick={ui_goto_edit_page}>Edit</div>
+	  <div title='Shortcut: FIXME' class='menu-button {$game.state.can_end_frame() ? "" : "unavailable"}' onclick={ui_end_frame}>End frame</div>
+	  <div title='Shortcut: FIXME' class='menu-button {$game.state.can_new_frame() ? "" : "unavailable"}' onclick={ui_new_frame}>New frame</div>
 	</div>
 	<div class='menu-column'>
-	  <div class='menu-button' on:click={ui_toggle_fullscreen}>Full screen</div>
-	  <div class='menu-button' on:click={ui_goto_start_page}>Main screen</div>
+	  <div class='menu-button' onclick={ui_toggle_fullscreen}>Full screen</div>
+	  <div class='menu-button' onclick={ui_goto_start_page}>Main screen</div>
 	  <div class='menu-button unavailable'>Help</div>
 	  <div class='menu-button unavailable'>Free ball</div>
 	</div>
